# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Event Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /events:
    get:
      operationId: listEvents
      tags: [events]
      summary: List events with filters
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 12
        - name: category
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: location
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: dateFrom
          in: query
          schema:
            type: string
        - name: dateTo
          in: query
          schema:
            type: string
        - name: sortBy
          in: query
          schema:
            type: string
            default: date
        - name: sortOrder
          in: query
          schema:
            type: string
            default: asc
      responses:
        "200":
          description: List of events
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventListResponse"
    post:
      operationId: createEvent
      tags: [events]
      summary: Create an event (Organizer/Admin)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventInput"
      responses:
        "201":
          description: Created event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
  /events/recommendations:
    get:
      operationId: getEventRecommendations
      tags: [events]
      summary: Get recommended events for current user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Recommended events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Event"
  /events/{id}:
    get:
      operationId: getEvent
      tags: [events]
      summary: Get event by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventDetail"
    patch:
      operationId: updateEvent
      tags: [events]
      summary: Update an event
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventUpdate"
      responses:
        "200":
          description: Updated event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
    delete:
      operationId: deleteEvent
      tags: [events]
      summary: Delete an event
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /events/{id}/status:
    patch:
      operationId: updateEventStatus
      tags: [events]
      summary: Approve or reject event (Admin)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventStatusUpdate"
      responses:
        "200":
          description: Updated event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
  /events/{id}/attendees:
    get:
      operationId: getEventAttendees
      tags: [events]
      summary: Get event attendees
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Attendee list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttendeeListResponse"
  /events/{id}/analytics:
    get:
      operationId: getEventAnalytics
      tags: [events]
      summary: Get event analytics
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Event analytics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventAnalytics"
...
  /organizer/stats:
    get:
      operationId: getOrganizerStats
      tags: [events]
      summary: Get organizer dashboard stats
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Organizer stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrganizerStats"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Event:
      type: object
      required: [id, title, description, date, venue, category, capacity, registrationDeadline, status, approvalStatus, organizerId, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        date:
          type: string
        venue:
          type: string
        category:
          type: string
        capacity:
          type: integer
        registrationDeadline:
          type: string
        status:
          type: string
          enum: [upcoming, ongoing, completed]
        approvalStatus:
          type: string
          enum: [pending, approved, rejected]
        bannerUrl:
          type: ["string", "null"]
        organizerId:
          type: integer
        organizerName:
          type: ["string", "null"]
        registrationCount:
          type: integer
        isRegistered:
          type: boolean
        createdAt:
          type: string
    EventDetail:
      type: object
      required: [id, title, description, date, venue, category, capacity, registrationDeadline, status, approvalStatus, organizerId, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        date:
          type: string
        venue:
          type: string
        category:
          type: string
        capacity:
          type: integer
        registrationDeadline:
          type: string
        status:
          type: string
          enum: [upcoming, ongoing, completed]
        approvalStatus:
          type: string
          enum: [pending, approved, rejected]
        bannerUrl:
          type: ["string", "null"]
        organizerId:
          type: integer
        organizerName:
          type: ["string", "null"]
        organizerEmail:
          type: ["string", "null"]
        registrationCount:
          type: integer
        isRegistered:
          type: boolean
        registrationId:
          type: ["integer", "null"]
        createdAt:
          type: string
    EventInput:
      type: object
      required: [title, description, date, venue, category, capacity, registrationDeadline]
      properties:
        title:
          type: string
        description:
          type: string
        date:
          type: string
        venue:
          type: string
        category:
          type: string
        capacity:
          type: integer
        registrationDeadline:
          type: string
        bannerUrl:
          type: string
    EventUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        date:
          type: string
        venue:
          type: string
        category:
          type: string
        capacity:
          type: integer
        registrationDeadline:
          type: string
        bannerUrl:
          type: string
        status:
          type: string
    EventStatusUpdate:
      type: object
      required: [approvalStatus]
      properties:
        approvalStatus:
          type: string
          enum: [approved, rejected]
        rejectionReason:
          type: string
    EventListResponse:
      type: object
      required: [events, total, page, totalPages]
      properties:
        events:
          type: array
          items:
            $ref: "#/components/schemas/Event"
        total:
          type: integer
        page:
          type: integer
        totalPages:
          type: integer
    AttendeeListResponse:
      type: object
      required: [attendees, total, page, totalPages]
      properties:
        attendees:
          type: array
          items:
            $ref: "#/components/schemas/Attendee"
        total:
          type: integer
        page:
          type: integer
        totalPages:
          type: integer
    Attendee:
      type: object
      required: [registrationId, userId, name, email, registeredAt, attended]
      properties:
        registrationId:
          type: integer
        userId:
          type: integer
        name:
          type: string
        email:
          type: string
        registeredAt:
          type: string
        attended:
          type: boolean
    EventAnalytics:
      type: object
      required: [eventId, totalRegistrations, capacity, attendanceRate, registrationsByDay]
      properties:
        eventId:
          type: integer
        totalRegistrations:
          type: integer
        capacity:
          type: integer
        attendanceRate:
          type: number
        registrationsByDay:
          type: array
          items:
            $ref: "#/components/schemas/DailyCount"
        categoryBreakdown:
          type: array
          items:
            $ref: "#/components/schemas/CategoryCount"
    DailyCount:
      type: object
      required: [date, count]
      properties:
        date:
          type: string
        count:
          type: integer
    CategoryCount:
      type: object
      required: [category, count]
      properties:
        category:
          type: string
        count:
          type: integer
...
    OrganizerStats:
      type: object
      required: [totalEvents, totalRegistrations, upcomingEvents, completedEvents]
      properties:
        totalEvents:
          type: integer
        totalRegistrations:
          type: integer
        upcomingEvents:
          type: integer
        completedEvents:
          type: integer
        pendingApproval:
          type: integer
        topEvents:
          type: array
          items:
            $ref: "#/components/schemas/EventSummary"
    EventSummary:
      type: object
      required: [id, title, registrationCount, capacity]
      properties:
        id:
          type: integer
        title:
          type: string
        registrationCount:
          type: integer
        capacity:
          type: integer