# lib/api-zod/src/generated/types/eventSummary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Event Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface EventSummary {
  id: number;
  title: string;
  registrationCount: number;
  capacity: number;
}